scripts/requirements.txt
reportlab
qrcode
Pillow
requests
python-dotenv
psycopg2-binary
